==> R/0-setup_project.R <==
#setup the package and working directories 

wd <- list()
setwd('~/NAPS-integrated-data/')

wd$data   <- paste0(getwd(),"/data/")
wd$output   <- paste0(getwd(),"/output/")
wd$header   <- paste0(getwd(),"/header_files/")
wd$R   <- paste0(getwd(),"/R/")
